Expose firebase authentication module from admin SDK
